+++ packages/pdf-viewer/src/plugins/PDFTooltipPlugin/TooltipCreator/TooltipCreator.tsx
@@ -5,23 +5,16 @@ import type { TooltipDataType } from '../def';
 
 const ef = () => {};
 
-export interface ShapeCreatorProps
-  extends React.HTMLAttributes<HTMLDivElement> {
+export interface ShapeCreatorProps extends React.HTMLAttributes<HTMLDivElement> {
   onDrawing?: (shape: TooltipDataType) => void;
   onCreate?: (shape: TooltipDataType) => void;
   onCancel?: () => void;
   pointMapping?: (point: { x: number; y: number }) => { x: number; y: number };
+  initialAttr?: Omit<Partial<TooltipDataType>, 'type'>;
 }
 
 const TooltipCreator = (props: ShapeCreatorProps) => {
-  const {
-    className = '',
-    onCreate = ef,
-    onDrawing = ef,
-    onCancel,
-    pointMapping = a => a,
-    ...otherProps
-  } = props;
+  const { className = '', onCreate = ef, onDrawing = ef, onCancel, pointMapping = a => a, initialAttr = {}, ...otherProps } = props;
   const styles = useStyle();
 
   const rootRef = useRef<HTMLDivElement>(null);
@@ -43,13 +36,14 @@ const TooltipCreator = (props: ShapeCreatorProps) => {
       y: startPosition[1] + offset[1],
     });
     _this.data = {
-      type: 'tooltip',
       width: 160,
       x: x2 + 16,
       y: y2 - 16,
       pointX: x1,
       pointY: y1,
       value: '请输入内容',
+      ...initialAttr,
+      type: 'tooltip',
     };
 
     if (finish) {
@@ -60,13 +54,7 @@ const TooltipCreator = (props: ShapeCreatorProps) => {
     }
   }, rootRef);
 
-  return (
-    <div
-      ref={rootRef}
-      className={`${styles.root} ${className}`}
-      {...otherProps}
-    />
-  );
+  return <div ref={rootRef} className={`${styles.root} ${className}`} {...otherProps} />;
 };
 
 export default TooltipCreator;