+++ packages/pdf-viewer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @orca-fe/pdf-viewer
 
+## 1.4.3
+
+### Patch Changes
+
+- feat: Painter，drawing 受控，增加 TooltipPlugin 的初始值設置
+- Updated dependencies
+  - @orca-fe/painter@1.2.3
+
 ## 1.4.2
 
 ### Patch Changes